backend/auth-service/requirements.txt
Flask
pymongo
redis
flask-bcrypt
PyJWT